Fix `subkey`/`subpath` nesting lifetimes instead of using the single `Hive` lifetime.
This removes all occurrences of the `H` type parameter.
It was meant to abstract over mutability, but did not allow to reuse a stored `Hive` reference without reborrowing (and introducing a sub-lifetime).
This commit uses a `Hive` reference with the single lifetime `'h` everywhere.

This wasn't possible without introducing `KeyNodeMut` next to `KeyNode` and duplicating some code, as we can no longer abstract over mutability.
However, this trade-off is worth it.